+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/dsl/PortForwardable.java
@@ -15,10 +15,14 @@
  */
 package io.fabric8.kubernetes.client.dsl;
 
+import io.fabric8.kubernetes.client.LocalPortForward;
+import io.fabric8.kubernetes.client.PortForward;
+
 import java.net.InetAddress;
+import java.nio.channels.ReadableByteChannel;
+import java.nio.channels.WritableByteChannel;
 
-public interface PortForwardable<H, L, I, O> {
-
+public interface PortForwardable {
   /**
    * Connects directly the remote port to the given input and output channels.
    *
@@ -27,16 +31,17 @@ public interface PortForwardable<H, L, I, O> {
    * @param out output channel
    * @return returns port forward interface
    */
-  H portForward(int port, I in, O out);
+  PortForward portForward(int port, ReadableByteChannel in, WritableByteChannel out);
 
   /**
-   * Create a server socket on a specified local port. Every connection to the local port will be forwarded to the remote port on the resource.
+   * Create a server socket on a specified local port. Every connection to the local port will be forwarded to the remote port
+   * on the resource.
    *
    * @param port port in integer
    * @param localPort local port in integer
    * @return returns local port forward interface
    */
-  L portForward(int port, int localPort);
+  LocalPortForward portForward(int port, int localPort);
 
   /**
    * Listen on port <code>localPort</code> on selected IP <code>inetAddress</code>, forwarding to <code>port</code>
@@ -47,15 +52,16 @@ public interface PortForwardable<H, L, I, O> {
    * @param localPort local port in integer
    * @return PortForward interface
    */
-  L portForward(int port, InetAddress localInetAddress, int localPort);
+  LocalPortForward portForward(int port, InetAddress localInetAddress, int localPort);
 
   /**
-   * Create a server socket on a random local port. Every connection to the local port will be forwarded to the remote port on the resource.
+   * Create a server socket on a random local port. Every connection to the local port will be forwarded to the remote port on
+   * the resource.
    * The random local port can be retrieved from the returned object.
    *
    * @param port port as integer
    * @return returns local port
    */
-  L portForward(int port);
+  LocalPortForward portForward(int port);
 
 }